--- ApiHH.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2027
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiHH", "ApiHH\ApiHH.csproj", "{3F2F3703-9B33-4C86-AF32-53B170448066}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3F2F3703-9B33-4C86-AF32-53B170448066}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F2F3703-9B33-4C86-AF32-53B170448066}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F2F3703-9B33-4C86-AF32-53B170448066}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F2F3703-9B33-4C86-AF32-53B170448066}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1686564D-0D7C-4BB9-B33D-CFD0CA50C3F8}
-	EndGlobalSection
-EndGlobal
