--- Lab5/dch
@@ -1,1 +1,0 @@
-...
